+++ src/v1/kafka/producer.ts
@@ -1,6 +1,6 @@
-import { userStatus } from '../status.schema.js';
+import { userStatus } from '../sockets/status/status.schema.js';
 import { TOPICS, USER_STATUS_EVENTS } from './constants.js';
-import { producer } from '../../../../plugins/kafka.js';
+import { producer } from '../../plugins/kafka.js';
 
 export async function sendStatus(userId: number, status: userStatus) {
   console.log(`Sending status update for user ${userId}: ${status}`);
