+++ src/v1/sockets/status/status.namespace.ts
@@ -1,6 +1,6 @@
 import { Namespace } from 'socket.io';
 import { socketMiddleware } from '../utils/middleware.js';
-import { startConsumer } from './kafka/consumer.js';
+import { startConsumer } from '../../kafka/consumer.js';
 import { handleConnection } from './connection.handler.js';
 
 export default async function statusNamespace(namespace: Namespace) {
